requirements.txt
aiohttp==3.9.3
aiocache==0.12.2
coc.py==3.2.1
cryptography==44.0.1
bcrypt==4.0.1
diskcache==5.6.3
expiring-dict==1.1.0
fastapi==0.111.0
fastapi-analytics==1.2.1
fastapi-cache2==0.2.1
fastapi-mail==1.4.1
gunicorn==22.0.0
jinja2==3.1.2
matplotlib==3.8.2
motor==3.3.2
orjson==3.9.13
pandas==2.2.0
passlib==1.7.4
pendulum==3.0.0
pillow==10.2.0
pydantic==2.6.0
PyJWT==2.10.1
pymongo==4.6.1
python-dotenv==1.0.1
python-dateutil==2.8.2
python-snappy==0.6.1
pytz==2023.4
redis==5.0.1
requests==2.32.3
slowapi==0.1.8
starlette==0.37.2
ujson==5.9.0
uvloop==0.19.0
uvicorn==0.30.1
numpy==1.26.4
openpyxl==3.1.3
sentry-sdk[fastapi]